+++ src/compas_model/elements/__init__.py
@@ -1,20 +1,44 @@
-from .element import Element
-from .element import Feature
 from .element import reset_computed
-from .block import BlockElement
+from .element import Element
 from .block import BlockFeature
+from .block import BlockElement
 from .block import BlockGeometry
-from .plate import PlateElement
 from .plate import PlateFeature
+from .plate import PlateElement
+from .column_head import ColumnHeadElement
+from .column_head import ColumnHeadCrossElement
+from .beam import BeamFeature
+from .beam import BeamElement
+from .beam import BeamIProfileElement
+from .beam import BeamSquareElement
+from .column import ColumnFeature
+from .column import ColumnElement
+from .column import ColumnRoundElement
+from .column import ColumnSquareElement
+from .fasteners import FastenersFeature
+from .fasteners import FastenersElement
+from .fasteners import ScrewElement
 
 
 __all__ = [
-    "Element",
-    "Feature",
-    "reset_computed",
-    "BlockElement",
-    "BlockFeature",
-    "BlockGeometry",
-    "PlateElement",
-    "PlateFeature",
+    reset_computed,
+    Element,
+    BlockFeature,
+    BlockElement,
+    BlockGeometry,
+    PlateFeature,
+    PlateElement,
+    ColumnHeadElement,
+    ColumnHeadCrossElement,
+    BeamFeature,
+    BeamElement,
+    BeamIProfileElement,
+    BeamSquareElement,
+    ColumnFeature,
+    ColumnElement,
+    ColumnRoundElement,
+    ColumnSquareElement,
+    FastenersFeature,
+    FastenersElement,
+    ScrewElement,
 ]