+++ src/compas_model/interactions/__init__.py
@@ -2,8 +2,10 @@ from .interaction import Interaction
 from .contact import (
     ContactInterface,
 )
+from .boolean_modifier import BooleanModifier
 
 __all__ = [
     "Interaction",
     "ContactInterface",
+    "BooleanModifier",
 ]